requirements.txt
discord~=1.7.3
python-dotenv~=0.18.0